[SPEC-6726] Move request interface into public and some cleanups
